conductor/test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# Hacking already pins down pep8, pyflakes and flake8
hacking>=2.0.0
# bandit>=1.1.0 # Apache-2.0
pytest>=5.4.1
coverage>=3.6 # Apache-2.0
fixtures>=3.0.0 # Apache-2.0/BSD
kombu>=3.0.25 # BSD
mock>=2.0 # BSD
mox3>=0.7.0 # Apache-2.0
oslotest>=1.10.0 # Apache-2.0
testrepository>=0.0.18 # Apache-2.0/BSD
testscenarios>=0.4 # Apache-2.0/BSD
testtools>=1.4.0 # MIT
os-testr>=1.0.0 # Apache-2.0
tempest>=11.0.0  # Apache-2.0
pifpaf>=0.0.11
junitxml>=0.7
requests-mock>=1.5.2
tosca-parser>=2.2.0

